# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DLU OneDesk API Contract
  description: Tài liệu chuẩn OpenAPI v3 cho toàn bộ RESTful API endpoints của hệ thống DLU OneDesk.
  version: 1.2.0
servers:
  - url: http://localhost:3000/api
    description: Local Development Server
paths:
...
  /tickets/guest:
    post:
      summary: Khách vãng lai gửi báo cáo sự cố qua QR Code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [studentId, title, description, deviceId]
              properties:
                studentId:
                  type: string
                title:
                  type: string
                description:
                  type: string
                deviceId:
                  type: string
      responses:
        "200":
          description: Tạo ticket thành công
        "429":
          description: Vượt quá giới hạn lượt gửi (Rate-limited)
...